# fragmenting relay 2/n: Update lazyCallReq to return header and args