--- a/relay_messages.go
+++ b/relay_messages.go
@@ -246,12 +246,6 @@ func (f *lazyCallReq) Arg2Iterator() (arg2.KeyValIterator, error) {
 	if !bytes.Equal(f.as, _tchanThriftValueBytes) {
 		return arg2.KeyValIterator{}, fmt.Errorf("non thrift scheme %s", f.as)
 	}
-
-	// TODO(echung): drop this since we already do error checks in the constructor
-	if f.arg2EndOffset > f.Header.PayloadSize() {
-		return arg2.KeyValIterator{}, errBadArg2Len
-	}
-
 	return arg2.NewKeyValIterator(f.Payload[f.arg2StartOffset:f.arg2EndOffset])
 }