--- a/relay_messages_test.go
+++ b/relay_messages_test.go
@@ -51,6 +51,7 @@ type testCallReqParams struct {
 	argScheme       Format
 	arg2Buf         []byte
 	overrideArg2Len int
+	arg3Buf         []byte
 }
 
 func (cr testCallReq) req(tb testing.TB) lazyCallReq {
@@ -113,6 +114,8 @@ func (cr testCallReq) frameWithParams(p testCallReqParams) *Frame {
 	}
 	payload.WriteUint16(uint16(arg2Len))
 	payload.WriteBytes(p.arg2Buf)
+	payload.WriteUint16(uint16(len(p.arg3Buf)))
+	payload.WriteBytes(p.arg3Buf)
 
 	return f
 }
@@ -307,6 +310,68 @@ func TestLazyCallReqSetTTL(t *testing.T) {
 	})
 }
 
+func byteKeyValToMap(tb testing.TB, buffer []byte) map[string]string {
+	rbuf := typed.NewReadBuffer(buffer)
+	nh := int(rbuf.ReadSingleByte())
+	retMap := make(map[string]string, nh)
+	for i := 0; i < nh; i++ {
+		keyLen := int(rbuf.ReadSingleByte())
+		key := rbuf.ReadBytes(keyLen)
+		valLen := int(rbuf.ReadSingleByte())
+		val := rbuf.ReadBytes(valLen)
+		retMap[string(key)] = string(val)
+	}
+	require.NoError(tb, rbuf.Err())
+	return retMap
+}
+
+func uint16KeyValToMap(tb testing.TB, buffer []byte) map[string]string {
+	rbuf := typed.NewReadBuffer(buffer)
+	nh := int(rbuf.ReadUint16())
+	retMap := make(map[string]string, nh)
+	for i := 0; i < nh; i++ {
+		keyLen := int(rbuf.ReadUint16())
+		key := rbuf.ReadBytes(keyLen)
+		valLen := int(rbuf.ReadUint16())
+		val := rbuf.ReadBytes(valLen)
+		retMap[string(key)] = string(val)
+	}
+	require.NoError(tb, rbuf.Err())
+	return retMap
+}
+
+func TestLazyCallReqContents(t *testing.T) {
+	cr := reqHasAll.reqWithParams(t, testCallReqParams{
+		arg2Buf: thriftarg2test.BuildKVBuffer(map[string]string{
+			"foo": "bar",
+			"baz": "qux",
+		}),
+		arg3Buf: []byte("some arg3 data"),
+	})
+
+	t.Run(".header()", func(t *testing.T) {
+		assert.Equal(t, map[string]string{
+			"cn":      "fake-caller",
+			"k1":      "v1",
+			"k222222": "",
+			"k3":      "thisisalonglongkey",
+			"rd":      "fake-delegate",
+			"rk":      "fake-routingkey",
+		}, byteKeyValToMap(t, cr.header()))
+	})
+
+	t.Run(".arg2()", func(t *testing.T) {
+		assert.Equal(t, map[string]string{
+			"baz": "qux",
+			"foo": "bar",
+		}, uint16KeyValToMap(t, cr.arg2()))
+	})
+
+	t.Run(".arg3()", func(t *testing.T) {
+		assert.Equal(t, "some arg3 data", string(cr.arg3()))
+	})
+}
+
 func TestLazyCallArg2Offset(t *testing.T) {
 	wantArg2Buf := []byte("test arg2 buf")
 	tests := []struct {
@@ -349,23 +414,44 @@ func TestLazyCallArg2Offset(t *testing.T) {
 	}
 
 	t.Run("no arg3 set", func(t *testing.T) {
-		for _, testHasMore := range []bool{true, false} {
-			t.Run(fmt.Sprintf("hasMore flag is set=%v", testHasMore), func(t *testing.T) {
+		tests := []struct {
+			msg       string
+			hasMore   bool
+			wantError string
+		}{
+			{
+				msg:     "hasMore flag=true",
+				hasMore: true,
+			},
+			{
+				msg:       "hasMore flag=false",
+				wantError: "buffer is too small",
+			},
+		}
+
+		for _, tt := range tests {
+			t.Run(fmt.Sprintf(tt.msg), func(t *testing.T) {
 				withLazyCallReqCombinations(func(crt testCallReq) {
 					// For each CallReq, we first get the remaining space left, and
 					// fill up the remaining space with arg2.
 					crNoArg2 := crt.req(t)
 					arg2Size := int(crNoArg2.Header.PayloadSize()) - crNoArg2.Arg2StartOffset()
 					var flags byte
-					if testHasMore {
+					if tt.hasMore {
 						flags |= hasMoreFragmentsFlag
 					}
-					cr := crt.reqWithParams(t, testCallReqParams{
+					f := crt.frameWithParams(testCallReqParams{
 						flags:   flags,
 						arg2Buf: make([]byte, arg2Size),
 					})
+					cr, err := newLazyCallReq(f)
+					if tt.wantError != "" {
+						require.EqualError(t, err, tt.wantError)
+						return
+					}
+					require.NoError(t, err)
 					endOffset, hasMore := cr.Arg2EndOffset()
-					assert.Equal(t, hasMore, testHasMore)
+					assert.Equal(t, hasMore, tt.hasMore)
 					assert.EqualValues(t, crNoArg2.Header.PayloadSize(), endOffset)
 				})
 			})