… typed/buffer_test.go