In order to mutate a frame and potentially do a fragmenting send, we'll need access to the original content of the frame, specifically the header and arguments. This change refactors `newLazyCallReq` to keep track of the offsets of its contents, and adds some methods for exposing the data.