@cinchurge 2020-08-11
before change: sizeof(lazyCallReq)=152
after change: sizeof(lazyCallReq)=144