@prashantv 2020-08-11
> benchmarks show no significant impact:
> 
> ```
> $ cat baseline.log | grep allocs/op
>    16287	     65357 ns/op	  15.67 MB/s	     443 B/op	      12 allocs/op
>    24693	     49499 ns/op	  20.69 MB/s	     478 B/op	      12 allocs/op
>    13987	     75435 ns/op	  54.30 MB/s	     469 B/op	      12 allocs/op
> $ cat benchmark.log | grep allocs/op
>    17516	     67423 ns/op	  15.19 MB/s	     415 B/op	      12 allocs/op
>    23822	     60139 ns/op	  17.03 MB/s	     471 B/op	      12 allocs/op
>    12994	     93586 ns/op	  43.77 MB/s	     454 B/op	      12 allocs/op
> ```

Hmm, every one of those benchmarks has lower (sometimes by a lot) MB/s -- can you run with a longer benchtime `-benchtime 5s` and run it a few times to see if it's consistent. Allocs are good, but seems like CPU is impacted more than I expected from this change.