@cinchurge 2020-08-11
> Hmm, every one of those benchmarks has lower (sometimes by a lot) MB/s -- can you run with a longer benchtime `-benchtime 5s` and run it a few times to see if it's consistent. Allocs are good, but seems like CPU is impacted more than I expected from this change.

```
   78856	     72196 ns/op	  14.18 MB/s	     407 B/op	      12 allocs/op
  116932	     53841 ns/op	  19.02 MB/s	     423 B/op	      12 allocs/op
   72776	     87787 ns/op	  46.66 MB/s	     405 B/op	      12 allocs/op

   79749	     76678 ns/op	  13.35 MB/s	     407 B/op	      12 allocs/op
  103214	     59519 ns/op	  17.20 MB/s	     426 B/op	      12 allocs/op
   64903	     89494 ns/op	  45.77 MB/s	     424 B/op	      12 allocs/op

   78600	     68604 ns/op	  14.93 MB/s	     406 B/op	      12 allocs/op
  118072	     50952 ns/op	  20.10 MB/s	     422 B/op	      12 allocs/op
   70366	     84699 ns/op	  48.36 MB/s	     409 B/op	      12 allocs/op

   86704	     68498 ns/op	  14.95 MB/s	     401 B/op	      12 allocs/op
  110632	     56660 ns/op	  18.07 MB/s	     425 B/op	      12 allocs/op
   72034	     81536 ns/op	  50.24 MB/s	     413 B/op	      12 allocs/op
```
yea it does seem that there is a consistent drop in MB/s, though not by a whole lot? should we be concerned?