@cinchurge 2020-08-13
the degradation appears to be related to the use of `byte` as an iteration variable, which was not necessary anyway. the benchmark results now appear to be comparable:
```
baseline:
   13946	     77751 ns/op	  52.68 MB/s	     469 B/op	      12 allocs/op
   13573	     85754 ns/op	  47.76 MB/s	     472 B/op	      12 allocs/op
   13183	     88546 ns/op	  46.26 MB/s	     459 B/op	      12 allocs/op

this change:
   14059	     76475 ns/op	  53.56 MB/s	     461 B/op	      12 allocs/op
   13468	     83500 ns/op	  49.05 MB/s	     446 B/op	      12 allocs/op
   13504	     86154 ns/op	  47.54 MB/s	     436 B/op	      12 allocs/op
```