@prashantv 2020-08-14
I think we should be running the benchmark in relay_messages_benchmark_test.go to benchmark the impact of this change.

For comparison,
master:
```
BenchmarkCallReqFrame   35954797               165 ns/op               0 B/op          0 allocs/op
```

dev:
```
BenchmarkCallReqFrame    3472144              1666 ns/op             592 B/op          5 allocs/op
```

This is just a regression caused by use of the `require` package in the benchmark, which allocates like crazy. Removing that,
```
BenchmarkCallReqFrame   20682429               251 ns/op             160 B/op          1 allocs/op
```

The allocation is from when we switched over to using pointers (expected), which likely is causing an increase in the overall time spent.

With this change, there's another slight increase:
```
BenchmarkCallReqFrame   21282790               267 ns/op             144 B/op          1 allocs/op
```

So overall, 165 -> 250 -> 267, which seems reasonable since we're doing more work.

Can you update the benchmark to get rid of require:
```
-               require.NoError(b, err)
+               if err != nil {
+                       b.Fatal("unexpected error", err)
+               }
```

And then post your pre-post benchmark numbers.