@cinchurge 2020-08-17
Benchmark results for newLazyCallReq:
dev:
```
BenchmarkCallReqFrame-8   	 5394974	       220 ns/op	     160 B/op	       1 allocs/op
```
this change:
```
BenchmarkCallReqFrame-8   	 5082253	       238 ns/op	     176 B/op	       1 allocs/op
```