@cinchurge 2020-08-17
> the degradation appears to be related to the use of `byte` as an iteration variable, which was not necessary anyway. the benchmark results now appear to be comparable:
> 
> ```
> baseline:
>    13946	     77751 ns/op	  52.68 MB/s	     469 B/op	      12 allocs/op
>    13573	     85754 ns/op	  47.76 MB/s	     472 B/op	      12 allocs/op
>    13183	     88546 ns/op	  46.26 MB/s	     459 B/op	      12 allocs/op
> 
> this change:
>    14059	     76475 ns/op	  53.56 MB/s	     461 B/op	      12 allocs/op
>    13468	     83500 ns/op	  49.05 MB/s	     446 B/op	      12 allocs/op
>    13504	     86154 ns/op	  47.54 MB/s	     436 B/op	      12 allocs/op
> ```

turns out this is probably just a red herring - redoing the benchmark now doesn't really show systematic difference:
using `int` to iterate:
```
   13083	     84326 ns/op	  12.14 MB/s	     430 B/op	      12 allocs/op
   17619	     64461 ns/op	  15.89 MB/s	     475 B/op	      12 allocs/op
   12234	    105773 ns/op	  38.72 MB/s	     451 B/op	      12 allocs/op
```
using `byte` to iterate:
```
   11590	    102040 ns/op	  10.04 MB/s	     435 B/op	      12 allocs/op
   17680	     68194 ns/op	  15.02 MB/s	     491 B/op	      12 allocs/op
   10999	    106316 ns/op	  38.53 MB/s	     480 B/op	      12 allocs/op
```