@cinchurge 2020-08-17
> Can you fix the lint issue causing one of the builds to fail:
> 
> > typed/buffer.go:95:1: exported method ReadBuffer.SkipBytes should have comment or be unexported
> 
> Let's also try and add coverage for the new code (ideally the patch coverage should be close to 100%)

yea, working on coverage right now